coup: Remove the check for is_finished when handling callback data
If is_finished is true when there is still something on the stack then
the game would end up getting stuck because nothing would allow the
game to progress and it doesn’t respond to the end game situation
until the bottom of the stack is reached.

This was particularily causing a problem with the exchange cards
action because that leaves the player’s cards in an intermediate
state. If the player’s second card was already dead then this would be
copied to the first card slot and the game would think the player is
dead which can cause is_finished to think the game is over.