zombie: Show all the dice in one message after 6 seconds
Otherwise the messages are very likely to hit the limit of 20 messages
per second. The 6 seconds is calculated because when the throw button is
hit it displays two messages. Which means if the player presses the
throw button instantly there will be a message every 3 seconds, which is
exactly the limit.