coup: Add allegiances
Instead of just choosing between the inspector and the ambassador, the
players can now also choose to play the reformation extension. In that
case all of the players are assigned an allegiance and it’s not possible
to attack a player of the same allegiance. Currently it’s not possible
to convert or embezzle.