message-queue: Use a rolling window to check the mpm
Instead of counting each minute period, it now has a rolling window and
counts the number of messages since the last message that was sent less
than a minute ago. With the previous approach it would have been
possible for example to send a message then send another 17 messages 59
seconds later. If another message is set 1 second later then it would be
considered a new period even though the previous 17 messages were only a
second ago.