superfight: Don’t keep the deciding fighter after a draw
After a decider fight, the winner should keep their original fighter,
not the temporary one. In order to implement this, the start_argument
message now takes a pointer to the fighters instead of always using the
main ones. start_decider_fight now creates the temporary fighters on the
stack instead of replacing the main ones.