game: Combine both send message callbacks to a super struct
Instead of having two send message callbacks with a bunch of arguments,
they are now combined into one and they take a pointer to a struct
instead of all the arguments. There is a define with the default values
for the struct properties. This makes it easy to add more properties
without having to touch all of the code.