pucobox: Add a lot more leeway to what is considered being scrolled down
On mobile devices, when the keyboard pops up the chat window ends up
being as if it was scrolled up by quite a lot and then it no longer
automatically scrolls to the bottom when a new message is received. This
tries to fix it by making the scroll amount 3/4 of the client height.