server: Add a language code to the new_player message
The language code is used as an additional criteria to find a matching
pending game. The game is then started with that language. The server
tries to filter out languages that haven’t been translated for the
WebSocket server and also for games that haven’t been translated into
that language.