web: Fix the layout of the title and help button
The previous approach made the page scroll horizontally on small screens
because it seems to take into account the size of the help button before
it is offset. This new approach uses a flex layout with a dummy element
on the left that is the same width as the button.