server: Change ref_pending_conversation to get_pending_conversation
When a pending conversation is added, a reference will always be held to
it in the list of pending conversations. Therefore there is no need to
add an extra reference to it when returning from the function. This will
make the reference counting a little easier to manage in later patches
when private conversations are added.