web: Use height:100% instead of 100vh
Otherwise on Chrome and the new version of Firefox this seems to end up
having a height that includes the region under the address bar. In
Chrome this was just a little bit annoying because it made the page
scroll, but in the new Firefox it breaks the chat because the address
bar never disappears and you can’t access the input box.