connection: Don’t call SSL_write with a zero-sized buffer
This situation can arrive as explained in the comment. This was making
SSL_write return 0 but with no error, and then Pucxobot was just closing
the connection. It is easily repeatable with Snitch because it sends a
private message just to the player who clicked the button when selecting
a role.