connection: Handle the remote end closing connection in SSL_read
If the remote end closes the connection without sending a WebSocket
close frame then it would previously try to report this as an SSL error
in the logs but the error would be 0 and the log looked confusing.
Instead it is now handled the same way as when read returns 0.