fox: Fix handling the 9 card
Previously I had misunderstood that the suit of the 9 card replaces the
trump suit. It’s actually supposed to be that the suit of the 9 card
changes to be the same as the suit of the trump card. This can make a
difference if the two cards played are in the same suit.

https://boardgamegeek.com/filepage/153033?commentid=10201066#comment10201066