proto: Fix writing the frame length on big-endian systems
Previously it was converting to BE to store the number and then the
storing function converts it back to LE. On both big-endian and
little-endian systems this ends up as doing one swap, which is wrong.
This changes it to always do the first swap so that on BE systems it
will do two swaps which is equivalent to doing nothing.

This patch is a hypothetical fix for BE systems which haven’t been
tested.