proto: Check that incoming strings are valid UTF-8
I think if a malicious client sent malformed UTF-8 it wouldn’t really
cause any harm except that the server would probably forward the invalid
strings to other clients and then they would get Javascript errors when
they try to decode them. If we validate the strings then it will avoid
this and also help if we start doing more text processing later on. For
the Telegram bot I don’t think it should be possible to have invalid
strings because the JSON parser would reject it.