wordparty: Fix deciding when to pick a new syllable after a fail
Previously when giving everyone the syllable, if one of the players lost
all of their lives before everyone received the syllable then it would
change too soon. To fix this it now stores the number of players that
were still in play when the syllable was chosen instead of recounting
every time it decides.