bot: Remove the buttons after choosing a game
When the bot asks which game the user wants to play, it previously left
the buttons there forever. It is really tempting to click on them and
flood the group with messages. What’s worse is that there’s no way to
know who clicked it. To fix this the bot now just removes the buttons
after one is clicked.

Fixes #15