game/conversation: Store sideband data instead of asking for it
Previously there was a problem where if a client reconnects after the
game had finished then it would try to ask the game for the messages for
the sideband data but the game would have already been destroyed and it
would crash. The mechanism was also a bit sloppy because the layer for
handling the websocket protocol leaked into the abstraction for
implementing a game. To fix this, instead of asking the game for a
protocol message to update the sideband data, the game now just directly
sets the data using the callback and the server part handles sending the
message with the data. Currently the data is just a single byte for each
data point. If more games start using the sideband data later then the
protocol could be modified again to store more types.