test-wordparty: Check that we get the word result again even if same
Wordparty is sort of misusing the sideband data to report an event when
sending a word result instead of just using it to report the state of
the game. This test is to guard against a future possible optimisation
that if the sideband data has the same value it won’t resend the data.
That change would break Wordparty.