web/pucxo: Add a mechanism to have a URL that jumps directly to the game
If the last part of the URL used to load the document is the title of
one of the games then it will now directly jump to that game instead of
having the player choose. The goal is to be able to have a URL to to
give to play Wordparty.