vortolisto: Only check the word type for variants
Instead of checking that the variant ends with exactly the same ending
after the root in the main head word, it now checks at the point of
determining what type of word it is by just looking at the ending. This
is needed for example to catch variants like “bezin·stacio” for
“bezin·ejo” because the part after the root is completely different. We
also can’t just analyse the variants independently because some of them
are just written out as a complete word without the <tld> tag.