vortlisto: If the variant contains a <tld> then handle it separately
The parse_kap function now returns None for the after part if the kap
doesn’t contain the <tld> tag. That way we can detect if one was found
in a variant and treat it like a normal kap instead of checking if it
has the same ending as the head word.